api/requirements.txt
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
Flask==1.1.2
gunicorn==20.0.4
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
mysql-connector-python==8.0.22
protobuf==3.14.0
python-dotenv==0.15.0
requests==2.24.0
six==1.15.0
spotipy==2.16.1
urllib3==1.25.11
Werkzeug==1.0.1
